+++ project/plugins/sitemap/routes/web.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Flextype;
-
-use Flextype\Plugin\Sitemap\Controllers\SitemapController;
-
-flextype()->get('/' . flextype('registry')->get('plugins.sitemap.settings.route'), SitemapController::class . ':index')->setName('sitemap.index');