+++ project/plugins/twig/CHANGELOG.md
@@ -1,3 +1,60 @@
+<a name="3.2.0"></a>
+# [3.2.0](https://github.com/flextype-plugins/twig/compare/v3.1.0...v3.2.0) (2022-09-13)
+
+### Features
+
+* **core** Add new functions `abs` and `round` for Math Extension.
+
+* **core** Add new filter `expressions`.
+
+* **core** Add new extensions `Fetch`.
+
+
+<a name="3.1.0"></a>
+# [3.1.0](https://github.com/flextype-plugins/twig/compare/v3.0.0...v3.1.0) (2022-09-13)
+
+### Features
+
+* **core** update code base for new Flextype 1.0.0-alpha.2 release.
+
+<a name="3.0.0"></a>
+# [3.0.0](https://github.com/flextype-plugins/twig/compare/v2.1.0...v3.0.0) (2022-07-11)
+
+### Features
+
+* **core** Updates code base for new Flextype 1.0.0-alpha.1 release.
+
+* **core** Added new Twig expression.
+
+* **core** Added new Twig shortcode.
+
+* **core** Added new Twig directive.
+
+* **core** Added new Twig extensions:
+  - I18n
+  - Filesystem
+  - Csrf
+  - Constants
+  - Strings
+  - Collection
+  - Filters
+  - Functions
+  - Tests
+  - Emitter
+  - Actions
+  - Registry
+  - Entries
+  - Serializers
+  - Parsers
+  - Flash
+  - Slugify
+  - Cache
+  - Image
+  - Url
+  - Math
+
+* **core** Updated and removed unused code.
+
 <a name="2.1.0"></a>
 # [2.1.0](https://github.com/flextype-plugins/twig/compare/v2.0.0...v2.1.0) (2021-01-14)